requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb==0.5.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4
sqlalchemy==2.0.31
pytest==8.2.2
httpx==0.27.0
numpy==1.26.4
